Reduce transaction scope and heartbeat frequency
connects #2500 on github

These two features together lead to a large increase in task runtime in
cases where multiple tasks were running in parallel. This is because
each task would hold a transaction with edits to one import_event record
for the entire life of the task. The result is that each task would have
to wait until a sibling task finishes its transaction. We were seeing
queries in excess of ten seconds as a result. By scoping the transaction to just the
length of the row edits, queries from other tasks need only block as long as it takes a
task to edit each row.

Additionally, the heartbeat was being run once per row (85 times per
task) when once per task is perfectly adequate.

It is suspected, but not confirmed, that the heartbeat is the particular
aspect of the row edit that was causing the lock on import_event that
was blocking other tasks. It is possible that performing either one of
these two changes would have fixed this issue. However, they are both
improvements so there is little benefit to further investigation.

To test:
1) run an import job with rate limiting at its current setting of
1/m. Note the time per job.
2) change rate limiting value to something like 15/m.
3) run an import job. If this fix works, the tasks run more frequently
but with the same average runtime.